+++ src/main/kotlin/com/fancychild/bapsanghead/controller/WebExceptionHandler.kt
@@ -1,7 +1,7 @@
 package com.fancychild.bapsanghead.controller
 
-import com.fancychild.bapsanghead.dto.response.ErrorResponse
-import com.fancychild.bapsanghead.dto.response.ErrorResponse.Companion.of
+import com.fancychild.bapsanghead.controller.dto.response.ErrorResponse
+import com.fancychild.bapsanghead.controller.dto.response.ErrorResponse.Companion.of
 import com.fancychild.bapsanghead.exception.BaseException
 import com.fancychild.bapsanghead.exception.ErrorCode
 import jakarta.servlet.http.HttpServletRequest
