+++ src/main/kotlin/com/fancychild/bapsanghead/controller/dto/request/RegisterRequest.kt
@@ -1,19 +1,19 @@
-package com.fancychild.bapsanghead.dto.request
+package com.fancychild.bapsanghead.controller.dto.request
 
 import io.swagger.v3.oas.annotations.media.Schema
 
-data class UpdateUserDetailsRequest(
+data class RegisterRequest(
         @Schema(description = "키", example = "180")
-        val height: Int? = null,
+        val height: Int,
 
         @Schema(description = "몸무게", example = "70")
-        val weight: Int? = null,
+        val weight: Int,
 
         @Schema(description = "나이", example = "25")
-        val age: Int? = null,
+        val age: Int,
 
         @Schema(description = "성별", example = "MALE or FEMALE")
-        val gender: String? = null,
+        val gender: String,
 
         @Schema(
                 description = "활동량(영어로 보내주세요.)",
@@ -22,5 +22,5 @@ data class UpdateUserDetailsRequest(
                         "or MEDIUM(활동적인 편입니다.) " +
                         "or HIGH(운동 없으면 못삽니다.)"
         )
-        val activityLevel: String? = null
+        val activityLevel: String
 )
