+++ src/main/kotlin/com/fancychild/bapsanghead/filter/JwtFilter.kt
@@ -1,7 +1,7 @@
 package com.fancychild.bapsanghead.filter
 
 import com.fancychild.bapsanghead.domain.user.dto.UserDto
-import com.fancychild.bapsanghead.dto.response.ErrorResponse
+import com.fancychild.bapsanghead.controller.dto.response.ErrorResponse
 import com.fancychild.bapsanghead.exception.BaseException
 import com.fancychild.bapsanghead.exception.ErrorCode
 import com.fancychild.bapsanghead.jwt.JwtAuthTokenUtil
